ajout confirmation suppression

+++ src/app/components/admin-product/admin-product.component.ts
@@ -3,6 +3,7 @@ import { Product } from 'src/app/models/product';
 import { MenuService } from 'src/app/services/menu.service';
 import { Router } from '@angular/router';
 
+
 @Component({
   selector: 'app-admin-product',
   templateUrl: './admin-product.component.html',
@@ -11,6 +12,8 @@ import { Router } from '@angular/router';
 export class AdminProductComponent implements OnInit {
 
   products: Product[];
+  searchText: string;
+
 
   constructor(private menuService: MenuService, private router:Router) { }
 
@@ -33,29 +36,44 @@ export class AdminProductComponent implements OnInit {
     this.router.navigate(['add-product']);
   }
 
-
 //https://www.devglan.com/angular/angular-6-example
   deleteProduct(product : Product) {
-
-    console.log("deleted");
-    this.menuService.deleteProducts(product.ID).subscribe(
-      data => {
-        this.products = data;
-        //je récupère les nouvelles data
-      })
+//ouverture de la confirmation de suppression grace à angular Material
+//https://www.youtube.com/watch?v=L7mrAYsh0-0
+    this.menuService.openDialog().afterClosed().subscribe(
+        result => {
+          //je supprime seulement si confirmer est cliqué
+          if(result == false){
+            this.menuService.deleteProducts(product.ID).subscribe(
+              data => {
+                this.products = data;
+                //je récupère les nouvelles data
+              })
+          }
+        }
+    );
   }
 
   editProduct(productID:number){
     this.router.navigate(['edit-product', productID]);
   }
 
+  //Otherwise you can't click on every part of the button
   handleCategory(){
     this.router.navigate(['admin-category']);
   }
-
 
   handleSubCategory(){
     this.router.navigate(['admin-sub-category']);
   }
 
-}
+
+
+
+
+
+
+  }
+  
+
+
